+++ src/main/java/com/gmail/val59000mc/threads/TeleportPlayersThread.java
@@ -8,7 +8,6 @@ import com.gmail.val59000mc.players.UhcTeam;
 
 import java.util.logging.Logger;
 
-import org.bukkit.Location;
 import org.bukkit.entity.Player;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
@@ -44,9 +43,7 @@ public class TeleportPlayersThread implements Runnable{
 
 			uhcPlayer.freezePlayer(team.getStartingLocation());
 
-			// Add 2 blocks to the Y location to prevent players from spawning underground.
-			Location location = team.getStartingLocation().clone().add(0, 2, 0);
-			player.teleport(location);
+			player.teleport(team.getStartingLocation());
 
 			player.removePotionEffect(PotionEffectType.BLINDNESS);
 			player.setFireTicks(0);