+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - [#29][GL-29] - Teams are sometimes not correctly updated on the tab list
 - [#45][GL-45] - Errors are spammed to the server console on Minecraft 1.19.3+ servers if ProtocolLib is installed
 - [#55][GL-55] - Display names are not obfuscated in the Anonymous scenario if use-team-colors is set to false
+- [#56][GL-56] - Players may sometimes fall into the ground (and suffocate) when teleported to their starting location
 
 [GL-10]: https://gitlab.com/uhccore/uhccore/-/issues/10
 [GL-18]: https://gitlab.com/uhccore/uhccore/-/issues/18
@@ -37,6 +38,7 @@
 [GL-45]: https://gitlab.com/uhccore/uhccore/-/issues/45
 [GL-47]: https://gitlab.com/uhccore/uhccore/-/issues/47
 [GL-55]: https://gitlab.com/uhccore/uhccore/-/issues/55
+[GL-56]: https://gitlab.com/uhccore/uhccore/-/issues/56
 
 ## [1.20.3] - 2022-12-17